requirements.txt
-i https://pypi.org/simple
alembic==1.11.1 ; python_version >= '3.7'
aniso8601==9.0.1
bcrypt==4.0.1 ; python_version >= '3.6'
blinker==1.6.2 ; python_version >= '3.7'
click==8.1.4 ; python_version >= '3.7'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
email-validator==2.0.0.post2
faker==18.13.0
flask==2.3.2
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.4
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
honcho==1.1.0
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.8.0 ; python_version < '3.10'
importlib-resources==6.0.0 ; python_version < '3.9'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
psycopg2-binary==2.9.6
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
setuptools==68.0.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.18 ; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
typing-extensions==4.7.1 ; python_version >= '3.7'
werkzeug==2.3.6 ; python_version >= '3.8'
wtforms==3.0.1 ; python_version >= '3.7'
zipp==3.16.0 ; python_version >= '3.8'
